`reduce()`: Reducer always receives an index
The index may be ignored by the reducer, but `reduce()` still provides
it. If it's bound to a function parameter, the parameter's value will never be
`undefined`; it will always be a `number`.

This mirrors `Array.prototype.map()`, and is already reflected in the first
overload of `reduce()`, so I believe this was just a mistake.